third_party/blink/web_tests/external/wpt/editing/run/formatblock_1001-2000-expected.txt

This is a testharness.js-based test.
Found 24 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<del>[foobar]</del>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<quasit>[foobar]</quasit>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p> foo</p>" but got "<p></p><p> </p><p>foo</p>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p> foo</p>" but got "<p></p><p> </p><p>foo</p>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>foo</p></td><td><p>bar</p></td><td><p>baz</p></td></tr></tbody></table>" but got "<p><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></p>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>foo</p></td><td><p>bar</p></td><td><p>baz</p></td></tr></tbody></table>" but got "<p><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></p>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>foo</p></td><td><p>bar</p></td><td><p>baz</p></td></tr></tbody></table>" but got "<p><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></p>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>foo</p></td><td><p>bar</p></td><td><p>baz</p></td></tr></tbody></table>" but got "<p><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></p>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<dl><dt>[foo]<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><dl><dd>bar</dd></dl>" but got "<dl><p>foo</p><dd>bar</dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<dl><dt>[foo]<dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><dl><dd>bar</dd></dl>" but got "<dl><p>foo</p><dd>bar</dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<dl><dt>foo<dd>[bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt></dl><p>bar</p>" but got "<dl><dt>foo</dt><p>bar</p></dl>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<dl><dt>foo<dd>[bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt></dl><p>bar</p>" but got "<dl><dt>foo</dt><p>bar</p></dl>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<dl><dt>[foo<dd>bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "<dl><p>foo<br>bar</p></dl>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<dl><dt>[foo<dd>bar]</dl>" queryCommandIndeterm("formatblock") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<dl><dt>[foo<dd>bar]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "<dl><p>foo<br>bar</p></dl>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<dl><dt>[foo<dd>bar]</dl>" queryCommandIndeterm("formatblock") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<ol><li>[foobar]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foobar</p></li></ol>" but got "<p><ol><li>foobar</li></ol></p>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<ol><li>[foobar]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foobar</p></li></ol>" but got "<p><ol><li>foobar</li></ol></p>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<ul><li>[foobar]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li><p>foobar</p></li></ul>" but got "<p><ul><li>foobar</li></ul></p>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<ul><li>[foobar]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li><p>foobar</p></li></ul>" but got "<p><ul><li>foobar</li></ul></p>"
Harness: the test ran to completion.
